+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [v0.1.1.0](https://github.com/freckle/amazonka-mtl/compare/v0.1.0.0...v0.1.1.0)
 
-- Add `withAuth` and `modified`
+- Add `withAuth` and `localEnv`
 
 ## [v0.1.0.0](https://github.com/freckle/amazonka-mtl/tree/v0.1.0.0)